Update initialization method
Now sched_wrr_entity is initalized in a function. The reason for doing this is it could be
called from various places such as enqueue_task_wrr. Real time scheduler did not use a
function to do this but fields of sched_rt_entity are changed. I think we can call this
function and do whatever we need to change the fields afterward.

Also fixes style.