Set sched_class in rt_mutex_prio()
This _may_ be the reason we have been freezing our system. When tested
by priority RT>FAIR>WRR we don't freeze any more. But the task seems to
have never be requeued. When doing `while (1) { print_pid(); }`, only
one pid is printed. The task status is always RUNNING.

In addition to the above problem, We should test this by:
 - expected priority
 - kthread & init's descedants' policies set to WRR