Fix style of last commit
The hrtimer_restart function does not hold any lock because I assume only
timer interrupt will touch it. This may not be correct. We need to refer
to examples in the kernel. This version compiles but I did not boot and
test it. Waiting for Andreas to finish task migration.